gratitude_journal_api/requirements.txt
fastapi
uvicorn
python-multipart
authlib
python-jose[cryptography]
pydantic
redis
passlib[bcrypt]